requirements.txt
asyncpg
aiogram
APScheduler
python-decouple
SQLAlchemy
psycopg2
torch
opencv-python
ultralytics